+++ XcodeML-Exc-Tools/src/exc/object/UnionType.java
@@ -8,7 +8,7 @@ package exc.object;
 
 public class UnionType extends CompositeType
 {
-    public UnionType(String id, XobjList id_list, int typeQualFlags, Xobject gccAttrs)
+    public UnionType(String id, XobjList id_list, long typeQualFlags, Xobject gccAttrs)
     {
         super(Xtype.UNION, id, id_list, typeQualFlags, gccAttrs);
     }
